instances blueberry_honey and strawberry_honey were NOT merged together
Query: in the Honey class, should the instances 'Blueberry Honey' and 'Strawberry Honey' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	 No.

While both Blueberry Honey and Strawberry Honey are types of honey, they are not semantic or ontological duplicates because they have different flavors and properties that result from the type of flower nectar used by the bees to make them. Therefore, it is important for them to remain separate classes.
meaning that instances blueberry_honey and strawberry_honey are different.
Query cache in files:
- /kgfiller/data/cache-bb60449f16125a501c4cd6a2c87e070b554a3aa5a9986b26fcdec80fb06bf8ca.yml