instances eucalyptus_honey and honey were NOT merged together
Query: in the Honey class, should the instances 'Eucalyptus Honey' and 'Honey' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	 No.

While both Eucalyptus honey and regular honey are types of honey, they have distinct differences in terms of their source (the type of plant nectar they are made from) and flavor profile. Therefore, it would not be appropriate to merge them as semantic and ontological duplicates. Keeping these separate allows for more accurate categorization and differentiation between various kinds of honey.
meaning that instances eucalyptus_honey and honey are different.
Query cache in files:
- /kgfiller/data/cache-6d13683ec9dd2ecc723d3aaa1f50962f1fedbf44323f31ba81a698353ef96266.yml