instances acacia_honey and honey were NOT merged together
Query: in the Honey class, should the instances 'Acacia Honey' and 'Honey' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	 No.

While both Acacia honey and regular honey are types of honey, they have different properties that may be relevant when considering them from a dietary perspective. For example, Acacia honey has a milder taste and is lower on the glycemic index compared to other honeys. Therefore, it would not be appropriate to merge these two concepts as semantic and ontological duplicates.
meaning that instances acacia_honey and honey are different.
Query cache in files:
- /kgfiller/data/cache-7de8db3a1ef7170a85c38c57ddb3939bcff23025a842f37e885e66f71818ade4.yml